Stage 3: The Taker's behavior.
The Taker will shout loud in pain if it can't find a suitable Maker to
hit its #no of swaps target. There is no good way to pre-know this, as a
Maker can be malicious and drop connection even after saying "Hello" and
sending us his "Offer".

This is the actual malice situation and in such cases the Taker will abort, and
recover from all the ongoing swaps. It will first broadcast all "outgoing" and
"watching" contract txs. And the it will start its recover thread and wait for the
timelock to mature.

The other Makers on the route will see thir incoming contrac txs in the blockchain,
will immediately know something is wrong, and start their own recovery mechanism.